Move more common functionality into utils and bench_results
