+++ tools/performance/engine-benchmarks/website_regen.py
@@ -4,18 +4,17 @@ IMPORTANT NOTE: Should be run only on the CI!!
 This script regenerate the benchmark results website, hosted as GH web pages on the
 https://github.com/enso-org/engine-benchmark-results repo.
 """
-import asyncio
 import logging
 from argparse import ArgumentParser
 from datetime import datetime, timedelta
 from pathlib import Path
 from typing import List, Dict, Set
 
-from bench_tool import Source, JobRun, JobReport, GH_DATE_FORMAT, TemplateBenchData, JinjaData
-from bench_tool.bench_results import get_bench_runs, get_bench_report
+from bench_tool import Source, JobRun, JobReport, TemplateBenchData, JinjaData
+from bench_tool.bench_results import get_bench_runs, fetch_job_reports
 from bench_tool.remote_cache import SyncRemoteCache
 from bench_tool.template_render import create_template_data, render_html
-from bench_tool.utils import WithTempDir, gather_all_bench_labels
+from bench_tool.utils import gather_all_bench_labels, sort_job_reports
 
 # The inception date of the benchmarks, i.e., the date of the first benchmark run.
 ENGINE_SINCE = datetime.fromisoformat("2022-12-01")
@@ -55,31 +54,13 @@ async def generate_bench_website(
         )
     assert len(bench_runs) > 0, "No benchmark runs found"
 
-    job_reports: List[JobReport] = []
-
-    async def _process_report(_bench_run: JobRun):
-        with WithTempDir("website-regen") as temp_dir:
-            _job_report = await get_bench_report(_bench_run, temp_dir, remote_cache)
-        if _job_report:
-            job_reports.append(_job_report)
-
-    tasks = []
-    for bench_run in bench_runs:
-        tasks.append(_process_report(bench_run))
-    await asyncio.gather(*tasks)
-
+    job_reports = await fetch_job_reports(bench_runs, remote_cache)
     _logger.debug(f"Gathered {len(job_reports)} job reports")
     assert len(job_reports) > 0, "No job reports found"
 
     _logger.debug("Sorting job_reports by commit date")
+    sort_job_reports(job_reports)
 
-    def _get_timestamp(job_report: JobReport) -> datetime:
-        return datetime.strptime(
-            job_report.bench_run.head_commit.timestamp,
-            GH_DATE_FORMAT
-        )
-
-    job_reports.sort(key=lambda report: _get_timestamp(report))
     all_bench_labels: Set[str] = gather_all_bench_labels(job_reports)
     _logger.debug(f"Found {len(all_bench_labels)} unique benchmark labels")
